# src/openapi/openapi.yml
openapi: 3.0.0
info:
  title: Todo List APIs
  description: Power X BE CAPSTONE
  version: 1.0.0
servers:
  - url: "http://localhost:3001/"
    description: local(dev)



components:
  schemas:
    User:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
    Login:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    Todo:
      type: object
      required:
        - title
        - by
        - created_on
      properties:        
        title:
          type: string   
        by:
          type: string
        created_on:
          type: string
          format: date
        updated_by:
          type: string
        updated_on:
          type: string
          format: date
    indivtask:
      type: object
      required:
        - task_title
      properties:
        task_title:
          type: string   
        updated_by:
          type: string
        updated_on:
          type: string
          format: date
        isFinished_on:
          type: boolean
  
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
      
paths:
  /register:
    post:
      tags:
      - auth
      summary: Register a user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User' 
            example:
              username: 'malcom@gmail.com'   
              name: 'malcom'
              password: 'password1'    
      responses:
        200:
          description: OK
        400:
          description: Username already exists
  /login:
    post:
      tags:
      - auth
      summary: Login a  Registered user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login' 
            example:
                  email: malcom@gmail.com
                  password: password1  
      responses:
        200:
          description: OK
        400:
          description: Username already exists

  /todo:
    post:
      tags:
      - Todo
      summary: Post a New Todo List
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
            example:
              title: 'Example todo'
              by: 'Malcom'
              created_on: 2021-10-17
              updated_by: "Malcom"
              updated_on: 2021-10-17
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'        
    get:
      tags:
      - Todo
      summary: Get all todos by current user
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
                example:
                - id: 
                  title: 'todo list 1'
                  by: 'John'
                - id: 
                  title: 'todo list 2'
                  by: 'John'                  
        400:
          description: "this person has no list of todos"

  /todo/{todo_id}:
    get:
      tags:
      - Todo
      summary: get a todolist based on the specified todo_id
                    user has to be creator role or collaborator role
      parameters:
        - in: path
          name: todo_id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
              example:
                todo_id: 1
                title: 'Example todo'
                by: 'Malcom'
                uid: 2
                created_on: 2021-10-17
                updated_by: "Malcom"
                updated_on: 2021-10-17
                soft_delet: 'false'
        400:
          description: "Enter Valid id Number or List of todos not found"

        403:
          description: "Neither Creator nor collaborator"

    put:
      tags:
      - Todo
      summary: update a todolist based on the specified todo_id
                    user has to be creator role or collaborator role
      parameters:
        - in: path
          name: todo_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
            example:
              title: 'Example todo'
              updated_by: 'Malcom'
              updated_on: 2021-10-17
              soft_delete: 'false'

      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
              example:
                todo_id: 1
                title: 'Example todo'
                by: 'Malcom'
                uid: 2
                created_on: 2021-10-17
                updated_by: "Malcom"
                updated_on: 2021-10-17
                soft_delet: 'false'
        400:
          description: "Enter Valid id Number or List of todos not found"

        403:
          description: "Neither Creator nor collaborator"

    delete:
      tags:
      - Todo
      summary:  SOFT delete a todolist based on the specified todo_id
                    user has to be creator role or collaborator role
      parameters:
        - in: path
          name: todo_id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
              example:
                todo_id: 1
                title: 'Example todo'
                by: 'Malcom'
                uid: 2
                created_on: 2021-10-17
                updated_by: "Malcom"
                updated_on: 2021-10-17
                soft_delete: 'true'
        403:
          description: Neither Creator or Collaborator

        400:
          description: "Enter Valid id Number or List of todos not found"
   
        404:
          description: todo_id of  list not found

  /todos/share/{todo_id}:
    post:
      tags:
      - Todo
      summary:  Share todo list to another user
      parameters:
        - in: path
          name: todo_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
            example:
              email: 'Watson@gmail.com'
              name: 'watson'
              role: 'read-only'
      responses:
        200:
          description: Share Request Sent     
          
        403:
          description: Neither Creator or Collaborator

        400:
          description: "Enter Valid id Number or List of todos not found"
   
        404:
          description: todo_id of  list not found

  /indivtask/{todo_id}:
    post:
      tags:
      - Individual Task
      summary: Adding a task title to a existing todo list
      parameters:
        - in: path
          name: todo_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/indivtask'
            example:
              task_title: 'cut grass'
              updated_on: 20/10/2021
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/indivtask'
              example:
                task_id: 2
                todo_id: 2
                task_title: 'clear trash'
                updated_by: 'Malcom'
                updated_on: 2021-10-17
                soft_delete: 'false'
        403:
          description: Neither Creator or Collaborator

        400:
          description: "Enter Valid id Number or List of todos not found"

  /indivtask/{task_id}:
    put:
      tags:
      - Individual Task
      summary: update a task based on the specified task_id
                    user has to be creator role or collaborator role
      parameters:
        - in: path
          name: task_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/indivtask'
            example:
              task_title: "take trash out"
              updated_by: "Barry"
              updated_on: 2021-10-21
              isFinished: false

      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/indivtask'
              example:
                task_id: 2
                todo_id: 2
                task_title: 'take trash out'
                updated_by: 'Barry'
                updated_on: 2021-11-21
                isFinished: true
        400:
          description: "Enter Valid id Number"

        403:
          description: "Neither Creator nor collaborator"

    delete:
      tags:
      - Individual Task
      summary:  SOFT delete a task based on the specified task_id
                    user has to be creator role or collaborator role
      parameters:
        - in: path
          name: task_id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/indivtask'
              example:
                task_id: 3
                todo_id: 2
                task_title: 'cut grass'
                updated_by: 'Barry'
                updated_on: 2021-10-17
                soft_delete: true
        403:
          description: Neither Creator or Collaborator

        400:
          description: "Enter Valid id Number or List of todos not found"

   
  
   
           
   


  